Split seechange.js into multiple source files
