+++ webap/Makefile
@@ -1,7 +1,8 @@
 INSTALLDIR = install
 
 toinstall = seechange_webap.py baseview.py conductor.py updater.py rkauth_flask.py run_webap.sh \
-	static/seechange.css static/seechange.js static/seechange_start.js \
+	static/seechange.css static/seechange_ns.js static/seechange.js static/seechange_start.js \
+	static/provenancetags.js static/exposuresearch.js static/exposurelist.js static/exposure.js \
 	static/rkwebutil.js static/rkauth.js static/resetpasswd_start.js \
 	static/favicon.ico \
 	templates/base.html templates/seechange_webap.html
