+++ webap/static/seechange_ns.js
@@ -1,0 +1,9 @@
+// Namespace, which is the only thing exported
+
+var seechange = {};
+
+// utils
+
+seechange.nullorfixed = function( val, num ) { return val == null ? null : val.toFixed(num); }
+
+export { seechange };
